Move add modal into separate component
Includes new rotation checkboxes
